\name{datasets}
\alias{datasets}
\keyword{datasets}

\title{ Example Data Sets }
\description{
  Example data sets for use in MARSS vignettes for the \code{\link{MARSS-package}}. 
  
  \itemize{
  \item{\link{plankton} Plankton datasets: Lake WA plankton 32-year time series and Ives et al data from West Long Lake. }
  \item{\link{SalmonSurvCUI} Snake River spring/summer chinook survival indices. }
  \item{\link{isleRoyal} Isle Royale wolf and moose data with temperature and precipitation covariates. }
    \item{\link{population-count-data} A variety of fish, mammal and bird population count data sets. }
    \item{\link{loggerhead} Loggerhead turtle tracking (location) data from ARGOS tags. }
    \item{\link{harborSeal} Harbor seal survey data (haul out counts) from Oregon, Washington and California, USA. }
}
}

