\name{MARSSboot}
\alias{MARSSboot}
\title{ Bootstrap MARSS Parameter Estimates }
\description{
Creates bootstrap parameter estimates and simulated (or bootstrapped) data (if appropriate).  This is a base function in the \code{\link{MARSS-package}}.
}
\usage{
MARSSboot(MLEobj, nboot = 1000, 
  output = "parameters", sim = "parametric", 
  param.gen = "MLE", control = NULL, silent = FALSE)
}

\arguments{
  \item{MLEobj}{ An object of class \code{\link{marssMLE}}. Must have a \code{$par} element containing MLE parameter estimates. }
  \item{nboot}{ Number of bootstraps to perform. }
  \item{output}{ Output to be returned: "data", "parameters" or "all".} 
  \item{sim}{ Type of bootstrap: "parametric" or "innovations". See Details. }
  \item{param.gen}{ Parameter generation method: "hessian" or "MLE". } 
  \item{control}{ The options in \code{MLEobj$control} are used by default. If supplied here, must contain all of the following:
    \describe{
      \item{\code{max.iter}}{ Maximum number of EM iterations. } 
      \item{\code{tol}}{ Optional tolerance for log-likelihood change.  If log-likelihood decreases less than this amount relative to the previous iteration, the EM algorithm exits. } 
      \item{\code{allow.degen}}{ Whether to try setting Q or R elements to zero if they appear to be going to zero. }  
    } 
  }
  \item{silent}{ Suppresses printing of progress bar. }
}

\details{
Approximate confidence intervals (CIs) on the model parameters can be calculated by numerically estimating the Hessian matrix (the matrix of partial 2nd derivatives of the parameter estimates).  The Hessian CIs (\code{param.gen="hessian"}) are based on the asymptotic normality of ML estimates under a large-sample approximation. CIs that are not based on asymptotic theory can be calculated using parametric and non-parametric bootstrapping (\code{param.gen="MLE"}).  In this case, parameter estimates are generated by the ML estimates from each bootstrapped data set.  The MLE method (kem or BFGS) is determined by \code{MLEobj$method}.

Stoffer and Wall (1991) present an algorithm for generating CIs via a non-parametric bootstrap for state-space models (\code{sim = "innovations"}).  The basic idea is that the Kalman filter can be used to generate estimates of the residuals of the model fit.  These residuals are then standardized and resampled and used to generate bootstrapped data using the MARSS model and its maximum-likelihood parameter estimates.  One of the limitations of the Stoffer and Wall algorithm is that it cannot be used when there are missing data, unless all data at time t are missing.  An alternative approach is a parametric bootstrap (\code{sim = "parametric"}), in which the ML parameter estimates are used to produce bootstrapped data directly from the state-space model.
}
\value{
  A list with the following components:
  \item{boot.params}{ Matrix (number of params x nboot) of parameter estimates from the bootstrap. }
  \item{boot.data}{ Array (n x t x nboot) of simulated (or bootstrapped) data (if requested and appropriate). }
  \item{marss}{ The \code{\link{marssMODEL}} object (form="marss") that was passed in via \code{MLEobj$marss}. }
  \item{nboot}{ Number of bootstraps performed. }
  \item{output}{ Type of output returned.} 
  \item{sim}{ Type of bootstrap. }
  \item{param.gen}{ Parameter generation method: "hessian" or "KalmanEM".  } 
}
\references{ 
Holmes, E. E., E. J. Ward, and M. D. Scheuerell (2012) Analysis of multivariate time-series using the MARSS package. NOAA Fisheries, Northwest Fisheries Science
Center, 2725 Montlake Blvd E., Seattle, WA 98112   Type \code{RShowDoc("UserGuide",package="MARSS")} to open a copy.

Stoffer, D. S., and K. D. Wall. 1991. Bootstrapping state-space models: Gaussian maximum likelihood estimation and the Kalman filter. Journal of the American Statistical Association 86:1024-1033.

Cavanaugh, J. E., and R. H. Shumway. 1997. A bootstrap variant of AIC for state-space model selection. Statistica Sinica 7:473-496.
 }
\author{ 
  Eli Holmes and Eric Ward, NOAA, Seattle, USA.

  eli(dot)holmes(at)noaa(dot)gov, eric(dot)ward(at)noaa(dot)gov
}
\seealso{ 
  \code{\link{marssMLE}}  \code{\link{marssMODEL}}  \code{\link{MARSSaic}}  
}
\examples{
  #nboot is set low in these examples in order to run quickly
  #normally nboot would be >1000 at least
  dat <- t(kestrel)
  dat <- dat[2:3,]
  #maxit set low to speed up the example
  kem <- MARSS(dat, model=list(U="equal",Q=diag(.01,2)), 
     control=list(maxit=50))
  # bootstrap parameters from a Hessian matrix
  hess.list <- MARSSboot(kem, param.gen="hessian", nboot=4)

  # from resampling the innovations (no missing values allowed)
  boot.innov.list <- MARSSboot(kem, output="all", sim="innovations", nboot=4)

  # from a parametric bootstrap
  boot.list <- MARSSboot(kem, output="all", sim="parametric", nboot=4)
  
  # bootstrapped parameter estimates
  boot.list$boot.params
}
