% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARSS-methods.R
\name{MARSSfit}
\alias{MARSSfit}
\title{Generic for fitting MARSS models}
\usage{
MARSSfit(x, ...)
}
\arguments{
\item{x}{a \link{marssMLE} object.}

\item{...}{additional arguments for the fitting function}
}
\description{
Uses the method of a marssMLE class object. Will call a function such as \code{\link[=MARSSkem]{MARSSkem()}}, \code{\link[=MARSSoptim]{MARSSoptim()}} or \code{MARSStmb()} in the {marssTMB} package.
}
