\name{MARSSinfo}
\alias{MARSSinfo}
\title{ MARSS Error Messages and Warnings }
\description{
  Prints out more information for MARSS error messages and warnings.
}
\usage{
MARSSinfo(number)
}
\arguments{
  \item{number}{ An error or warning message number.    }
}

\value{
  A print out of information.
}
\author{ 
  Eli Holmes, NOAA, Seattle, USA.  
}
\examples{
# Show all the info options
MARSSinfo()
}
