\name{marssm-class}
\Rdversion{1.1}
\docType{class}
\alias{marssm-class}
\alias{print,marssm-method}
\alias{summary,marssm-method}

\title{Class "marssm"}
\description{marssm objects describe the multivariate autoregressive state space models used in the package \code{\link{MARSS-package}}.} 	
\section{Methods}{
  \describe{
    \item{print}{\code{signature(x = "marssm")}: ... }
    \item{summary}{\code{signature(object = "marssm")}: ... }
	 }
}
\author{ 
  Kellie Wills, NOAA, Seattle, USA.  
  kellie(dot)wills(at)noaa(dot)gov
}
\keyword{classes}
