\name{marssMLE-class}
\Rdversion{1.1}
\docType{class}
\alias{marssMLE-class}
\alias{print,marssMLE-method}
\alias{summary,marssMLE-method}

\title{Class "marssMLE"}
\description{marssMLE objects specify maximum-likelihood estimation of multivariate autoregressive state-space models in the package \code{\link{MARSS-package}}.}
\section{Methods}{
  \describe{
    \item{print}{\code{signature(x = "marssMLE")}: ... }
    \item{summary}{\code{signature(object = "marssMLE")}: ... }
	 }
}
\references{ ~put references to the literature/web site here ~ }
\author{ 
  Kellie Wills, NOAA, Seattle, USA.  
  kellie(dot)wills(at)noaa(dot)gov
}
\keyword{classes}
