\name{MARSSresiduals.tt1}
\alias{MARSSresiduals.tt1}

\title{ MARSS One-Step-Ahead Residuals }
\description{
  Calculates the standardized (or auxiliary) one-step-ahead residuals, aka the innovations residuals and their variance.  Not exported. Access this function with \code{residuals(object, conditioning="t-1")}.
}
\usage{
MARSSresiduals.tt1(object, method="SS", normalize=FALSE)
}
\arguments{
  \item{object}{ An object of class \code{\link{marssMLE}}.}
  \item{method}{ Algorithm to use. Currently only "SS". }
  \item{normalize}{ TRUE/FALSE }
}
\value{
A list with the following components  
  \item{residuals}{ The model residuals E(V(t)|y(1:t-1),Theta), where Theta is the set of model parameters.  Called the innovations. Residuals associated with missing data will appear as NA. }
  \item{var.residuals}{ The variance of the model residuals as a n x n x TT matrix. The variance exists for all t values including missing data. }
  \item{std.residuals}{ The Cholesky standardized model residuals as a n x TT matrix. This is \code{residuals} multiplied by the inverse of the Cholesky decomposition of \code{var.residuals}. }
  \item{mar.residuals}{ The marginal standardized model residuals as a n x TT matrix. This is \code{residuals} multiplied by the inverse of the diagonal matrix formed by the square-root of the diagonal of \code{var.residuals}. }
  \item{msg}{ Any warning messages. This will be printed unless Object$control$trace = -1 (suppress all error messages). }

}
\details{

This function returns the conditional expected value (mean) and variance of the model one-step-ahead residuals.  'conditional' means in this context, conditioned on the observed data up to time \eqn{t-1} and a set of parameters.  

\strong{model.residuals}

The model residuals \eqn{v_t} are the difference between the data and the predicted data at time t conditioned on the data up to time \eqn{t-1}:
\deqn{ v_t = y_t - Z x_t - a}
In a state-space model, \eqn{x_t} is stochastic and the model residuals are a random variable.  \eqn{y_t} is also stochastic, though often observed unlike \eqn{x_t}.  The model residual random variable is:
\deqn{ V_t= Y_t - Z X_t - a}
The unconditional mean and variance of \eqn{V_t} is 0 and R.  This function (\code{residuals(MLEobj)}) returns the \emph{conditional} mean and variance of \eqn{V_t} (conditioned on the data up to time \eqn{t-1}).

\code{model.residuals} returns
\deqn{ \mathbf{y}_t - \mathbf{Z}_t E(\mathbf{X}_t | \mathbf{y}^{(1),t-1}, \Theta) - \mathbf{a}_t }{y_t - Z_t E(X_t | y(1,t-1), \Theta) - a_t}.

\code{var.residuals} returned by the function is the conditional variance of the residuals. This is the variance of \eqn{V_t} conditioned on the data and the parameter set \eqn{\Theta}.  The unconditional variance (no data) would just be \eqn{\mathbf{R}_t}{R(t)} and the conditional variance is 
\deqn{ \mathbf{R}+\mathbf{Z}_t \mathbf{V}_t^{t-1} \mathbf{Z}_t^\top }{R + Z_t Vtt1 t(Z_t)}

\strong{Standardized residuals}

\code{std.residuals} are Cholesky standardized residuals. These are the residuals muliplied by the inverse of the Cholesky decomposition of the variance matrix of the residuals: \deqn{ \Sigma_t^{-1/2} \hat{v}_t}{ Sigma_t^{-1/2} hat{v}_t}. These residuals are uncorrelated. 

The interpretation of the Cholesky standardized residuals is not straight-forward when the Q and R variance-covariance matrices are non-diagonal.  The residuals which were generated by a non-diagonal variance-covariance matrices are transformed into orthogonal residuals in MVN(0,I) space.  For example, if v is 2x2 correlated errors with variance-covariance matrix R. The transformed residuals (from this function) for the i-th row of v is a combination of the row 1 effect and the row 1 effect plus the row 2 effect.  So in this case, row 2 of the transformed residuals would not be regarded as solely the row 2 residual but rather how different row 2 is from row 1, relative to expected.  If the errors are highly correlated, then the Cholesky standardized residuals can look rather non-intuitive.

\code{mar.residuals} are the marginal standardized residuals. These are the residuals muliplied by the inverse of the diagonal matrix formed from the square-root of the diagonal of the variance matrix of the residuals: \deqn{ dg(\Sigma_t)^{-1/2} \hat{v}_t}{ dg(Sigma_t)^{-1/2} hat{v}_t}, where 'dg(A)' is the square matrix formed from the diagonal of A, aka \code{diag(diag(A))}. These residuals will be correlated if the variance matrix is non-diagonal. 

\strong{Normalized residuals}

If \code{normalize=FALSE}, the unconditional variance of \eqn{V_t} and \eqn{W_t} are R and Q and the model is assumed to be written as
\deqn{ y_t = Z x_t + a + v_t}
\deqn{ x_t = B x_{t-1} + u + w_t}
If normalize=TRUE, the model is assumed to be written
\deqn{ y_t = Z x_t + a + Hv_t}
\deqn{ x_t = B x_{t-1} + u + Gw_t}
with the variance of \eqn{V_t} and \eqn{W_t} equal to I (identity).

\code{residuals.marssMLE} returns the residuals defined as in the first equations. To get the residuals defined as Harvey et al. (1998) define them (second equations), then use \code{normalize=TRUE}.  In that case the unconditional variance of residuals will be I instead of R and Q.  Note, that the `normalized' residuals are not the same as the `standardized' residuals.  In former, the unconditional residuals have a variance of I while in the latter it is the conditional residuals that have a variance of I.

}

\author{ 
  Eli Holmes, NOAA, Seattle, USA.  

  eli(dot)holmes(at)noaa(dot)gov
} 
\seealso{ \code{\link{MARSSresiduals.tT}}, \code{\link{fitted.marssMLE}}, \code{\link{plot.marssMLE}} }
\examples{
  dat <- t(harborSeal)
  dat <- dat[c(2,11),]
  MLEobj <- MARSS(dat)
  
  residuals(MLEobj, conditioning="t-1")$std.residuals
}
\references{
Holmes, E. E. 2014. Computation of standardized residuals for (MARSS) models. Technical Report. arXiv:1411.0045.  This report shows how to compute the residuals from the Kalman smoother output without using the Harvey et al. (1998) algorithm.  The variance of the \eqn{\hat{v}(t)}{hat{v}(t)} for the missing values can then be computed.
}

