\name{popWrap-class}
\Rdversion{1.1}
\docType{class}
\alias{popWrap-class}

\title{Class "popWrap"}
\description{popWrap objects are wrapper object containing specifications and options for estimation of multivariate 
autoregressive state-space models in the package \code{\link{MARSS-package}}.}
\author{ 
  Kellie Wills, NOAA, Seattle, USA.  
  kellie(dot)wills(at)noaa(dot)gov
}
\keyword{classes}
