\name{predict}
\alias{predict}
\alias{forecast}

\title{ predict and forecast MARSS MLE objects }
\description{
  See the following help files:
  \itemize{
  \item{\code{\link{predict.marssMLE}} Predict and forecast. }
  \item{\code{\link{forecast.marssMLE}} Forecast. Use \code{\link{predict.marssMLE}} to call with argument \code{h}. }
  \item{\code{\link{plot.marssPredict}} Plot a prediction or forecast. }   
  \item{\code{\link{autoplot.marssPredict}} Plot a prediction or forecast using ggplot2 package.  }
  \item{\code{\link{print.marssPredict}} Print prediction or forecast. If \code{h!=0}, i.e. forecast, only the forecast is printed but the \code{\link{marssPredict}} object (in \code{pred}) has all the historical time steps also.  }

}
}