\name{residuals.marssMLE}
\alias{residuals}
\alias{residuals.marssMLE}
\alias{residuals.MARSS}
\alias{residuals.marss}
\title{ MARSS Standardized Residuals }
\description{
  Calculates the standardized (or auxilliary) residuals sensu Harvey, Koopman and Penzer (1998).
}
\usage{
\method{residuals}{marssMLE}(object,..., Harvey=TRUE, normalize=FALSE)
}
\arguments{
  \item{object}{ An object of class \code{\link{marssMLE}}.}
  \item{...}{ Not used.}
  \item{Harvey}{ TRUE/FALSE }
  \item{normalize}{ TRUE/FALSE }
}
\details{
The model residuals are 
hatvt = y_t - Z_t E(X_t|y_1^T) - a_t
E(X_t|y_1^T) is the Kalman smoother estimate of x_t.
The variance of the model residuals is the variance of hatvt over Y.  

If Harvey=TRUE, the function uses the algorithm on page 112 of Harvey, Koopman and Penzer (1998) to compute the conditional residuals and variance of the conditional residual.  If Harvey=FALSE, the function uses the equations in the technical report Holmes (2014).  The difference in the algorithm only concerns the missing values.

If \code{Harvey=TRUE}, there will be no variance calculation for the missing y.  If \code{Harvey=FALSE}, the variance of missing y are computed (via Holmes 2014)  If \code{exact=FALSE}, the default, a `missing values corrected' version of these is returned (via Holmes 2014). The interpretation is that these missing values are left out in the expectation E(X_t|y_1^T), but you can still imagine that these data exist and you need the variance of their hatvt. For example, if you are doing a leave-one-out cross-validation, the data exist and you need their variance because you are going to compute some diagnostics using the left-out data.  For outlier diagnostics and shock detection, the variances for the missing values are not needed.

The state residuals are 
E(W_t|y_1^T) = E(X_t|y_1^T) - B_t E(X_t-1|y_1^T) - u_t
E(X_t|y_1^T) is the Kalman smoother estimate of x_t.
The state residuals always exist since the expected value of the states exist without data and will be identical with \code{Harvey=TRUE} or \code{Harvey=FALSE}.  Generally speaking, E(W_t|y_1^T) is not 0 even if there are missing data. For the unconditional expectations, E(X_t)=B E(X_t-1) + u_t but that is not true for the conditional expectations.

If \code{normalize=TRUE}, the residuals are normalized to have variance of 1 as in the algorithm shown in Harvey et al., but residuals.marssMLE returns the unnormalized residuals by default because that is typically what the 'residuals()' functions in R return.  

\code{residuals.marssMLE} will return the standardized residuals sensu Harvey et al. (1998) for you in \code{std.residuals} for outlier and shock detection.  These are the model and state residuals scaled by the inverse square root of the missing values corrected variance of the residuals.  Note the standardized model residuals are set to NA when there are missing data (if there is no data point, there is no model residual). The standardized state residuals however always exist since the expected value of the states exist without data.  

Note that interpretation of the standardized residuals is not straight-forward when the Q and R variance-covariance matrices are non-diagonal.  The residuals which were generated by a non-diagonal variance-covariance matrices are transformed into orthogonal residuals in MVN(0,I) space.  For example, if v is 2x2 correlated errors with variance-covariance matrix R. The transformed residuals (from this function) for the i-th row of v is a combination of the row 1 effect and the row 1 effect plus the row 2 effect.  So in this case, row 2 of the transformed residuals would not be regarded as solely the row 2 residual but rather how different row 2 is from row 1, relative to expected.  If the errors are highly correlated, then the transformed residuals can look rather non-intuitive.
}
\value{
A list with the following components  
  \item{model.residuals}{ The smoothed model residuals E(v(t)|y(1:T),Theta), where Theta is the set of model parameters.  Sometimes called the smoothations. This is different than the Kalman filter innovations which are E(v(t)|y(1:t-1),Theta).}
  \item{state.residuals}{ The smoothed stated residuals E(x(t)|y(1:T))-E(x(t)|E(x(t-1)|y(1:T))).}
  \item{residuals}{ The model residuals as a (n+m) x TT matrix with \code{model.residuals} on top and \code{state.residuals} below. \code{model.residuals} is hat(eta_t) on page 112 of Harvey, Koopman and Penzer (1998).  }
  \item{var.residuals}{ The variance of the model residuals as a (n+m) x (n+m) x TT matrix. This is var(hat(\code{model.residuals})).}
  \item{std.residuals}{ The standardized model residuals as a (n+m) x TT matrix. This is \code{residuals} divided by the square root of \code{var.residuals}---although the code is using the matrix equivalent of that equation.}
}
\author{ 
  Eli Holmes, NOAA, Seattle, USA.  

  eli(dot)holmes(at)noaa(dot)gov
} 
\seealso{ \code{\link{MARSSkem}} \code{\link{marssMLE}} }
\examples{
  dat = t(harborSeal)
  dat = dat[c(2,11),]
  MLEobj = MARSS(dat)
  
  #state residuals
  state.resids1=residuals(MLEobj)$state.residuals
  #this is the same as
  states=MLEobj$states
  Q=coef(MLEobj,type="matrix")$Q
  state.resids2=states[,2:30]-states[,1:29]-matrix(coef(MLEobj,type="matrix")$U,2,29)
  #normalize to 1
  state.resids2=solve(t(chol(Q)))%*%state.resids2
  #compare the two
  cbind(t(state.resids1[,-30]),t(state.resids2))

  #standardized (by variance) model & state residuals
  residuals(MLEobj)$std.residuals
}
\references{
Harvey, A., S. J. Koopman, and J. Penzer. 1998. Messy time series: a unified approach. Advances in Econometrics 13: 103-144  (see page 112-113).  Eqn 21 is the Kalman eqns.  Eqn 23 and 24 is the backward recursion to compute the smoothations.  This function uses the MARSSkf output for eqn 21 and then implements the backwards recursion in eqn 23 and eqn 24.  Pages 120-134 discuss the use of standardized residuals for outlier and structural break detection.

de Jong, P. and J. Penzer. 1998. Diagnosing shocks in time series. Journal of the American Statistical Association 93: 796–806.  This one shows the same equations; see eqn 6.  This paper mentions the scaling based on the inverse of the sqrt (chol) of the variance-covariance matrix for the residuals (model and state together).  This is in the right column, half-way down on page 800.

Koopman, S. J., N. Shephard, and J. A. Doornik. 1999. Statistical algorithms for models in state space using SsfPack 2.2. Econometrics Journal 2: 113-166. (see pages 147-148).

Harvey, A. and S. J. Koopman. 1992. Diagnostic checking of unobserved-components time series models. Journal of Business & Economic Statistics 4: 377-389.

Holmes, E. 2014. Standardized residuals for MARSS models. This report shows how to compute the residuals from the Kalman smoother output without using the Harvey et al. (1998) algorithm.  The variance of the hatvt for the missing values can then be computed.
}

