% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARSS-methods.R
\name{MARSSfit}
\alias{MARSSfit}
\title{Generic for fitting MARSS models}
\usage{
MARSSfit(x, ...)
}
\arguments{
\item{x}{a \link{marssMLE} object.}

\item{...}{additional arguments for the fitting function}
}
\description{
This is an internal function used by \code{\link[=MARSS]{MARSS()}}. It is not
intended for use by users but needs to be exported so
the marssTMB package can use it. Uses the method of a marssMLE class object.
Will call a function such as \code{\link[=MARSSkem]{MARSSkem()}}, \code{\link[=MARSSoptim]{MARSSoptim()}} in the
MARSS package or \code{MARSStmb()} in the marssTMB package.
}
