\name{marssMLE-class}
\Rdversion{1.1}
\docType{class}
\alias{marssMLE-class}
\alias{print,marssMLE-method}
\alias{summary,marssMLE-method}
\alias{coef,marssMLE-method}
\alias{residuals,marssMLE-method}
\alias{predict,marssMLE-method}
\alias{simulate,marssMLE-method}

\title{Class "marssMLE"}
\description{
\code{\link{marssMLE}} objects specify fitted multivariate autoregressive state-space models (maximum-likelihood) in the package \code{\link{MARSS-package}}.}
\section{Methods}{
  \describe{
    \item{print}{\code{signature(x = "marssMLE")}: ... }
    \item{summary}{\code{signature(object = "marssMLE")}: ... }
    \item{coef}{\code{signature(object = "marssMLE")}: ... }
    \item{predict}{\code{signature(object = "marssMLE")}: ... }
    \item{simulate}{\code{signature(object = "marssMLE")}: ... }
	 }
}
\author{ 
  Eli Holmes, NOAA, Seattle, USA
  
  eli(dot)holmes(at)noaa(dot)gov
  
  Kellie Wills, NOAA, Seattle, USA.  

  kellie(dot)wills(at)noaa(dot)gov
}
\keyword{classes}
